+++ src/lib/axios/requests/catalog/get-filtered-products.ts
@@ -1,16 +1,15 @@
 import { isAxiosError } from 'axios';
 
 import { envVariables } from '@/config/commerce-tools-api';
+import { apiInstance } from '@/lib/axios/axios-instances';
 
-import { apiInstance } from '../axios-instances';
+import { axiosErrorMsgSchema } from '../schemas/axios-error-msg.schema';
+import { type Product, getProductsResultSchema } from '../schemas/product-schema';
 import { ProductsRequestArguments } from './catalog-types';
-import { axiosErrorMsgSchema } from './schemas/axios-error-msg.schema';
-import { type Product, getProductsResultSchema } from './schemas/product-schema';
 
 export async function getFilteredProducts(offset = 0, BEARER_TOKEN: string, filters: string): Promise<Product[]> {
   const queryArgs: ProductsRequestArguments = { limit: 7, offset };
   const query = `/${envVariables.PROJECT_KEY}/product-projections/search?limit=${queryArgs.limit}&offset=${queryArgs.offset}&${filters}`;
-  // console.log(`complete query without encode: ${query}`);
 
   try {
     const getProductsResult = await apiInstance.get(query, {
